% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WID_clocks.R
\name{WID_clocks}
\alias{WID_clocks}
\title{WID clocks}
\usage{
WID_clocks(beta, tissue = NULL)
}
\arguments{
\item{beta}{Methylation beta matrix object}

\item{tissue}{Tissue type; if "breast", the epithelial-fibroblast-fat-immune reference will be used in the EpiDISH algorithm. Defaults to NULL}
}
\value{
Returns a list object with WID clocks.
}
\description{
THE WID_clocks function uses a methylation beta matrix and returns the WID general, epithelial and immune clocks.
It also returns the WID-relative-epithelial-age (WID_rea) and WID-relative-immune-age (WID_ria) (∆WID epithelial or WID immune and WID general clock).
}
\details{
Authors: Barrett J., Herzog C. & Kim. Y.-N. et al., 2021
}
